requirements.txt
#telegram bot
python-telegram-bot
python-dotenv
python-decouple
requests